VideoStreaming/requirements.txt
imutils==0.5.3
dlibs==19.21.0
opencv-python==4.4.0.44
scipy==1.5.2  
Flask==1.1.2



